Refactor SIWE logic
