+++ app/datasources/cache/redis.py
@@ -1,0 +1,10 @@
+from functools import cache
+
+from redis import Redis
+
+from ...config import settings
+
+
+@cache
+def get_redis() -> Redis:
+    return Redis.from_url(settings.REDIS_URL)
